refactor(timetable): split DayTimeTable into per-component files
src/components/talks/DayTimeTable/index.tsx had grown to ~510 LOC with
several internal components. Extract them into sibling files:

- TimeLabel.tsx       — time label cell
- SlotTrackHeader.tsx — track-name + hashtag-copy header
- SessionCard.tsx     — session detail card (with private TriangleBadge
                        and SessionTypeLabel)
- CellRenderer.tsx    — cell content discriminator (with private
                        BadgedShell and LabelText)

index.tsx now contains only the DayTimeTable container, time-slot /
boundary derivation, and the desktop/mobile grid layout (221 LOC).
No behavior change.
