+++ src/components/talks/DayTimeTable/SlotTrackHeader.tsx
@@ -1,0 +1,43 @@
+"use client";
+
+import { Copy } from "lucide-react";
+import { useState } from "react";
+import { Button } from "@/components/ui/button";
+import { TRACK_STYLE } from "@/constants/timetable";
+import type { Track } from "@/types/timetable-api";
+
+export function SlotTrackHeader({ track }: { track: Track }) {
+  const [copySuccess, setCopySuccess] = useState(false);
+  const style = TRACK_STYLE[track.id];
+
+  const copyToClipboard = async (text: string) => {
+    try {
+      await navigator.clipboard.writeText(text);
+      setCopySuccess(true);
+      setTimeout(() => setCopySuccess(false), 2000);
+    } catch (err) {
+      console.error("クリップボードへのコピーに失敗しました", err);
+    }
+  };
+
+  return (
+    <div className={`${style.bg} p-2 text-center`}>
+      <span className={`font-bold ${style.text}`}>{track.name}</span>
+      <div className="flex justify-center mt-2">
+        <Button
+          onClick={() => copyToClipboard(track.hashtag)}
+          className="rounded-full bg-white text-black hover:bg-gray-100 px-4 py-1 text-sm font-medium h-auto flex items-center gap-2"
+        >
+          {copySuccess ? (
+            <span>コピーしました！</span>
+          ) : (
+            <>
+              <span>{track.hashtag}</span>
+              <Copy className="h-4 w-4" />
+            </>
+          )}
+        </Button>
+      </div>
+    </div>
+  );
+}
