+++ src/components/talks/DayTimeTable/index.tsx
@@ -1,302 +1,15 @@
 "use client";
 
-// import { AddToMyTimetableButton } from "@/components/talks/AddToMyTimetableButton";
-import { Copy, ExternalLink } from "lucide-react";
-import Link from "next/link";
-import { useMemo, useRef, useState } from "react";
-import { ProfileImage } from "@/components/talks/FallbackImage";
+import { useMemo, useRef } from "react";
 import { Button } from "@/components/ui/button";
-import { getSessionMasterBySessionId } from "@/constants/sessionMaster";
-import { TALK_TYPE, TRACK_KEYS, TRACK_STYLE } from "@/constants/timetable";
+import { TRACK_KEYS } from "@/constants/timetable";
 import { useTimetable } from "@/hooks/useTimetable";
 import { cn } from "@/lib/utils";
-import type {
-  Cell,
-  SessionContent,
-  TimeSlot,
-  TimetableResponse,
-  Track,
-  TrackKey,
-} from "@/types/timetable-api";
+import type { Cell, TimeSlot, TimetableResponse } from "@/types/timetable-api";
 import { myTimetable } from "@/utils/myTimetable";
-
-function SlotTrackHeader({ track }: { track: Track }) {
-  const [copySuccess, setCopySuccess] = useState(false);
-  const style = TRACK_STYLE[track.id];
-
-  const copyToClipboard = async (text: string) => {
-    try {
-      await navigator.clipboard.writeText(text);
-      setCopySuccess(true);
-      setTimeout(() => setCopySuccess(false), 2000);
-    } catch (err) {
-      console.error("クリップボードへのコピーに失敗しました", err);
-    }
-  };
-
-  return (
-    <div className={`${style.bg} p-2 text-center`}>
-      <span className={`font-bold ${style.text}`}>{track.name}</span>
-      <div className="flex justify-center mt-2">
-        <Button
-          onClick={() => copyToClipboard(track.hashtag)}
-          className="rounded-full bg-white text-black hover:bg-gray-100 px-4 py-1 text-sm font-medium h-auto flex items-center gap-2"
-        >
-          {copySuccess ? (
-            <span>コピーしました！</span>
-          ) : (
-            <>
-              <span>{track.hashtag}</span>
-              <Copy className="h-4 w-4" />
-            </>
-          )}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function TimeLabel({
-  timeText,
-  isActive = false,
-}: {
-  timeText: string;
-  isActive?: boolean;
-}) {
-  return (
-    <div
-      className={`relative flex flex-col gap-2 p-2 items-center justify-center text-center w-full h-full md:w-[99px] lg:w-[125px] ${
-        isActive
-          ? "bg-orange-200 after:content-[''] after:absolute after:top-0 after:right-0 after:w-[5px] after:h-full after:bg-orange-500 after:opacity-50"
-          : "bg-yellow-200"
-      }`}
-    >
-      <p
-        className={`text-sm lg:text-base font-bold ${
-          isActive ? "text-orange-500" : "text-yellow-700"
-        }`}
-      >
-        {timeText}
-      </p>
-      {isActive && <p className="text-xs text-orange-500 font-normal">now</p>}
-    </div>
-  );
-}
-
-function TriangleBadge({ cssVar }: { cssVar: string }) {
-  return (
-    <div
-      className="hidden md:block"
-      style={{
-        width: 0,
-        height: 0,
-        borderStyle: "solid",
-        borderWidth: "0 24px 24px 0",
-        borderColor: `transparent ${cssVar} transparent transparent`,
-        position: "absolute",
-        top: 0,
-        right: 0,
-      }}
-    />
-  );
-}
-
-function SessionTypeLabel({
-  sessionType,
-}: {
-  sessionType: SessionContent["sessionType"];
-}) {
-  const { name, color } = TALK_TYPE[sessionType];
-  return (
-    <span
-      className="inline-block rounded px-2 py-0.5 text-xs font-bold text-white"
-      style={{ backgroundColor: color }}
-    >
-      {name}
-    </span>
-  );
-}
-
-function LabelText({ label, link }: { label: string; link?: string }) {
-  if (!link) return <>{label}</>;
-  return (
-    <a
-      href={link}
-      target="_blank"
-      rel="noopener noreferrer"
-      className="underline hover:text-blue-purple-500 inline-flex items-center gap-1"
-    >
-      <span>{label}</span>
-      <ExternalLink className="h-4 w-4" aria-hidden="true" />
-    </a>
-  );
-}
-
-function SessionCard({
-  content,
-  trackKey,
-  trackName,
-  id,
-}: {
-  content: SessionContent;
-  trackKey: TrackKey;
-  trackName: string;
-  id?: string;
-}) {
-  const style = TRACK_STYLE[trackKey];
-  return (
-    <div
-      id={id}
-      className="bg-white px-5 pt-10 pb-4 md:py-5 min-h-32 h-full flex flex-col gap-2 items-start justify-start text-black-700 relative"
-    >
-      <div
-        className={cn(
-          style.bg,
-          style.text,
-          "block md:hidden py-1 px-2 absolute top-0 left-0 text-xs font-bold",
-        )}
-      >
-        {trackName}
-      </div>
-      <TriangleBadge cssVar={style.cssVar} />
-      <div className="flex items-center justify-between w-full">
-        <SessionTypeLabel sessionType={content.sessionType} />
-        {/* <AddToMyTimetableButton
-          talkId={content.sessions[0].id}
-          talkIds={content.sessions.map((s) => s.id)}
-          withCheckbox
-        /> */}
-      </div>
-      <div className="flex flex-col gap-5">
-        {content.sessions.map((ref) => {
-          const master = getSessionMasterBySessionId(ref.id);
-          const title = master?.title ?? "";
-          const speakerName = master?.speaker.name ?? "";
-          return (
-            <div key={ref.id} className="flex flex-col gap-1">
-              <Link
-                href={`/talks/${ref.id}`}
-                className="underline hover:text-blue-purple-500"
-              >
-                <p className="text-[16px]">{title}</p>
-              </Link>
-              <div className="flex items-center gap-2">
-                <span className="text-[14px]">{speakerName}</span>
-                <div className="relative h-6 w-6 rounded-full shrink-0 overflow-hidden">
-                  <ProfileImage
-                    speakerName={speakerName}
-                    profileImageUrl={master?.speaker.profileImageUrl}
-                  />
-                </div>
-              </div>
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-function BadgedShell({
-  variant,
-  isSingleTrack,
-  trackKey,
-  trackName,
-  children,
-}: {
-  variant: "closed" | "card";
-  isSingleTrack: boolean;
-  trackKey: TrackKey;
-  trackName: string;
-  children: React.ReactNode;
-}) {
-  const style = TRACK_STYLE[trackKey];
-  const sizeCls =
-    variant === "closed"
-      ? "bg-gray-50 min-h-16"
-      : "bg-white pt-10 pb-4 md:py-5 min-h-32";
-  return (
-    <div
-      className={cn(
-        sizeCls,
-        "px-5 h-full flex flex-col gap-2 items-center justify-center text-black-700 relative",
-      )}
-    >
-      {isSingleTrack && (
-        <div
-          className={cn(
-            style.bg,
-            style.text,
-            "block md:hidden py-1 px-2 absolute top-0 left-0 text-xs font-bold",
-          )}
-        >
-          {trackName}
-        </div>
-      )}
-      {children}
-    </div>
-  );
-}
-
-function CellRenderer({
-  cell,
-  trackNames,
-  id,
-}: {
-  cell: Cell;
-  trackNames: Record<string, string>;
-  id?: string;
-}) {
-  const isSingleTrack = cell.tracks.length === 1;
-  const headTrack = cell.tracks[0];
-  const trackName = trackNames[headTrack] ?? headTrack;
-  const c = cell.content;
-
-  if (c.type === "closed") {
-    return (
-      <BadgedShell
-        variant="closed"
-        isSingleTrack={isSingleTrack}
-        trackKey={headTrack}
-        trackName={trackName}
-      >
-        クローズ
-      </BadgedShell>
-    );
-  }
-
-  if (c.type === "labeled" && c.muted) {
-    return (
-      <div className="bg-gray-50 px-5 h-full min-h-16 flex items-center justify-center text-black-700">
-        <LabelText label={c.label} link={c.link} />
-      </div>
-    );
-  }
-
-  // labeled (non-muted) と displayLabel 付き session は同じ「白カード」シェル
-  if (c.type === "labeled" || c.displayLabel !== undefined) {
-    const label = c.type === "labeled" ? c.label : c.displayLabel;
-    return (
-      <BadgedShell
-        variant="card"
-        isSingleTrack={isSingleTrack}
-        trackKey={headTrack}
-        trackName={trackName}
-      >
-        <LabelText label={label ?? ""} link={c.link} />
-      </BadgedShell>
-    );
-  }
-
-  return (
-    <SessionCard
-      content={c}
-      trackKey={headTrack}
-      trackName={trackName}
-      id={id}
-    />
-  );
-}
+import { CellRenderer } from "./CellRenderer";
+import { SlotTrackHeader } from "./SlotTrackHeader";
+import { TimeLabel } from "./TimeLabel";
 
 const cellKeyOf = (cell: Cell) => `${cell.startTime}-${cell.tracks[0]}`;
 
@@ -366,9 +79,7 @@ export function DayTimeTable({ data }: { data: TimetableResponse }) {
     return candidates[0] ? cellKeyOf(candidates[0]) : null;
   }, [data.cells]);
 
-  // Mobile uses single-column flow per timeSlot.
-  // Cells starting at this slot.startTime are rendered in track order; cells
-  // continuing from earlier slots are skipped (already rendered).
+  // モバイルでは同一 startTime のセルを縦に並べるための索引。
   const cellsByStartTime = useMemo(() => {
     const m = new Map<number, Cell[]>();
     for (const c of data.cells) {
